requirements.txt
numpy<2.0.0
setproctitle
opencv-python
smbus
systemd
Flask
loguru
gunicorn
pillow